✨ feat: add get reviewer survey route
Adds GET /survey/:surveyUuid/:reviewerUuid. Warning: currently does not
use proper types, which are declared on Github and on the frontend. This
is a proof of concept to make sure that I wasn't asking Atulya & Olivia
to do something impossible with this ticket. And to prove to myself that
NestJS is workable and I'm just being overdramatic.